refactor(modularization): Phase 3 Step 5b3 — document_service move to KB domain + DI via sibling + 6 envelope schemas carve-out (#16)
Step 5b3 in the Phase 3 #1629 atomic sequence (per @符炫炜
msg=da460d8c 5b sub-ordering; document_service parity with 5b2b per
PM msg=848c7f76 follow-up direction).

**document_service move (aperag/service → KB domain)**:
- `git mv` (~1500 LOC, body preserved; git rewrites as new+modify
  due to full-file import overhaul).
- G1 boundary pass: `aperag.db.models` / `aperag.schema.view_models` /
  `aperag.service.*` imports rewritten to `aperag.domains.knowledge_base.*`
  + `aperag.domains.indexing.db.models` + `aperag.schema.common`.
- Consumer-owned Protocol DI: `marketplace_service` /
  `quota_service` reached through `_get_marketplace_ops()` /
  `_get_quota_ops()` imported from sibling
  `aperag.domains.knowledge_base.service.collection_service`. Both
  KB services share the wire-up, so `aperag/app.py` startup (Step
  5b2c) continues to wire one set of instances that both services
  see.
- Inline banned imports eliminated: raw `UserQuota` SQL query
  replaced with `_get_quota_ops().get_user_quotas(user)` lookup;
  `DocumentIndexType` / `random_id` lifted to module-level imports
  (from `aperag.domains.indexing.db.models` and
  `aperag.domains.knowledge_base.db.models._random_id` respectively).

**QuotaOps Protocol extension** (ports.py):
- Added `get_user_quotas(user_id) -> dict` surface (already
  implemented on legacy `quota_service`, no provider change
  required). Docstring explains the admin-controlled cap rationale
  for read-only separate-session lookup.

**6 envelope schemas carve-out** (kb.schemas + view_models):
- `UploadDocumentResponse`, `FailedDocument`,
  `ConfirmDocumentsResponse`, `FetchUrlResultItem`, `FetchUrlResponse`,
  `StagedDocumentsResponse` moved from `aperag.schema.view_models`
  to `aperag.domains.knowledge_base.schemas`. They belong to the KB
  document public surface; document_service builds them as response
  envelopes. `ConfirmDocumentsRequest` and `FetchUrlRequest` stay
  in view_models — they are FastAPI route inputs and migrate with
  `aperag.views.documents_v2` in Step 5a.
- `_bind_view_models_reexports` hook + view_models try-block
  re-import both updated so pre-migration callers continue to see
  the same class objects (Step 4b dual-hook pattern).

**legacy shim** (aperag/service/document_service.py):
- Plain re-export (`from ... import *` + explicit `DocumentService`
  + `document_service` singleton); zero runtime side effects since
  DI wire-up lives in `aperag/app.py` (5b2c) and document_service
  imports DI accessors from the sibling collection_service module.

**Verification**:
- `tests/unit_test/test_modularization_boundaries.py`: 13/13 green;
  G1 AST walk finds zero legacy-aggregate imports in
  `aperag/domains/knowledge_base/service/document_service.py`
  (including the previously-banned inline closure imports).
- Full unit suite: 670 passed / 28 skipped (one formerly-skipped
  test now collects because the import path is valid).
- `uvx ruff check` + `uvx ruff format --check` clean across 269
  files.
- Semantic smoke: `import aperag.service.document_service` returns
  the same `DocumentService` class object as the domain path;
  `aperag.app` startup wires DI for both services simultaneously.
